+++ OpenRA.Mods.Common/Traits/Buildable.cs
@@ -10,7 +10,7 @@
 
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Traits
 {
 	public class BuildableInfo : TraitInfo<Buildable>
 	{
