+++ OpenRA.Mods.Common/Traits/Player/ProvidesCustomPrerequisite.cs
@@ -13,7 +13,7 @@ using System.Linq;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Traits
 {
 	public class ProvidesCustomPrerequisiteInfo : ITraitInfo
 	{
