+++ OpenRA.Mods.RA/Activities/Transform.cs
@@ -10,6 +10,7 @@
 
 using System.Linq;
 using OpenRA.Mods.Common;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
