+++ OpenRA.Mods.RA/CrateAction.cs
@@ -10,6 +10,7 @@
 
 using System.Linq;
 using OpenRA.Mods.Common.Effects;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA
