+++ OpenRA.Mods.RA/Player/ClassicProductionQueue.cs
@@ -12,6 +12,7 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Mods.Common;
 using OpenRA.Mods.Common.Power;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
