+++ OpenRA.Mods.RA/Player/PlaceBuilding.cs
@@ -10,6 +10,7 @@
 
 using System.Linq;
 using OpenRA.Effects;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Mods.RA.Traits;
 using OpenRA.Primitives;
