+++ OpenRA.Mods.RA/Player/ProductionQueue.cs
@@ -13,6 +13,7 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Mods.Common;
 using OpenRA.Mods.Common.Power;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Traits;
