+++ OpenRA.Mods.RA/Transforms.cs
@@ -9,6 +9,7 @@
 #endregion
 
 using System.Collections.Generic;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 using OpenRA.Mods.RA.Render;
 using OpenRA.Mods.RA.Buildings;
