+++ OpenRA.Mods.RA/UtilityCommands/ActorStatsExport.cs
@@ -14,6 +14,7 @@ using System.Data;
 using System.Linq;
 using OpenRA.GameRules;
 using OpenRA.Mods.Common;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA.UtilityCommands
