+++ OpenRA.Mods.RA/Widgets/Logic/ProductionTooltipLogic.cs
@@ -13,6 +13,7 @@ using System.Drawing;
 using System.Linq;
 using OpenRA.Mods.Common;
 using OpenRA.Mods.Common.Power;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 using OpenRA.Widgets;
